lib/syscall_shim: Don't inline syscall enter/exit tab wrapper
Inlining the syscall enter/exit tables wrapper used for native system
calls leads to an increase of ~30KB in size of images as well as
an increase in icache and iTLB misses. Therefore, do not inline it and
instead turn it into a regular internal function.

With this occasion, rename vars.c to tabs.c since it is more suggestive
to it containing only syscall entry/exit tables related code.
