--- lib/syscall_shim/vars.c
@@ -1,9 +1,0 @@
-/* SPDX-License-Identifier: BSD-3-Clause */
-/* Copyright (c) 2024, Unikraft GmbH and The Unikraft Authors.
- * Licensed under the BSD-3-Clause License (the "License").
- * You may not use this file except in compliance with the License.
- */
-
-#include <uk/syscall.h>
-
-__thread unsigned long uk_syscall_nested_depth;
